pyproject.toml
[tool.poetry]
name = "bmad-mcp"
version = "0.1.0"
description = "A fused multi-agent framework for n8n workflow development."
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.1.0"
langchain-openai = "^0.1.1"
python-dotenv = "^1.0.0"
pyyaml = "^6.0.1"
requests = "^2.31.0"
anthropic = "^0.25.0"
langchain-anthropic = "^0.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"